requirements.txt
flask>=2.2.3
werkzeug>=2.2.3
line-bot-sdk>=2.4.2
openai>=1.3.0
python-dotenv>=1.0.0
redis>=4.6.0
flask-limiter>=3.3.1
mysql-connector-python>=8.1.0
waitress>=2.1.2
gunicorn>=21.2.0
APScheduler>=3.10.1
requests>=2.31.0
tiktoken>=0.5.1

# Testing Dependencies
pytest>=8.4.1
